requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-docx==0.8.11
Pillow==9.5.0
fpdf==1.7.2
urllib3==2.0.4
sqlalchemy psycopg2-binary
